requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
# anthropic==0.34.2
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
requests